+++ test/e2e/servers/trigger.spec.ts
@@ -201,7 +201,9 @@ describe("onchainBlockEventhandler", () => {
     expect(lnInvoice).not.toBeInstanceOf(Error)
 
     const { paymentRequest: request } = lnInvoice as LnInvoice
-    await pay({ lnd: lndOutside1, request })
+    pay({ lnd: lndOutside1, request })
+
+    await sleep(250)
 
     const hash = getHash(request)
     const invoice = await getInvoice({ id: hash, lnd: lnd1 })
